chars are brought back if too far away for 4s
