+++ 3scape/Assets/Scenes/Tutorial.unity
@@ -1688,6 +1688,7 @@ GameObject:
   m_Component:
   - component: {fileID: 155895182}
   - component: {fileID: 155895181}
+  - component: {fileID: 155895183}
   m_Layer: 0
   m_Name: Virtual Camera
   m_TagString: Untagged
@@ -1713,7 +1714,7 @@ MonoBehaviour:
   m_Priority: 1
   m_StandbyUpdate: 1
   m_LookAt: {fileID: 0}
-  m_Follow: {fileID: 382437317}
+  m_Follow: {fileID: 1472064910}
   m_Lens:
     FieldOfView: 40
     OrthographicSize: 4
@@ -1745,6 +1746,21 @@ Transform:
   m_Father: {fileID: 487621374}
   m_RootOrder: 0
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
+--- !u!114 &155895183
+MonoBehaviour:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 155895180}
+  m_Enabled: 1
+  m_EditorHideFlags: 0
+  m_Script: {fileID: 11500000, guid: 2a467833147517248931f913e25ee760, type: 3}
+  m_Name: 
+  m_EditorClassIdentifier: 
+  playerAveragePosition: {fileID: 1472064910}
+  knight: {fileID: 382437317}
+  mage: {fileID: 1170955590}
+  archer: {fileID: 833718569}
 --- !u!1 &161845132
 GameObject:
   m_ObjectHideFlags: 0
@@ -26154,7 +26170,7 @@ Transform:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 487621371}
   m_LocalRotation: {x: 0, y: 0, z: 0, w: 1}
-  m_LocalPosition: {x: 10.567538, y: -6, z: -9.8}
+  m_LocalPosition: {x: 9.627541, y: -6.83, z: -9.8}
   m_LocalScale: {x: 1, y: 1, z: 1}
   m_Children:
   - {fileID: 155895182}
@@ -27801,6 +27817,7 @@ Transform:
   - {fileID: 382437317}
   - {fileID: 833718569}
   - {fileID: 1170955590}
+  - {fileID: 1472064910}
   m_Father: {fileID: 0}
   m_RootOrder: 7
   m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
@@ -27876,7 +27893,7 @@ Prefab:
     - target: {fileID: 212140600227288710, guid: 588d570a52e35c5458973a7d067a2875,
         type: 2}
       propertyPath: m_SortingLayer
-      value: 4
+      value: 5
       objectReference: {fileID: 0}
     - target: {fileID: 1726022557058710, guid: 588d570a52e35c5458973a7d067a2875, type: 2}
       propertyPath: m_Name
@@ -30703,17 +30720,17 @@ MonoBehaviour:
   m_XDamping: 1
   m_YDamping: 1
   m_ZDamping: 1
-  m_ScreenX: 0.5
+  m_ScreenX: 0.4999999
   m_ScreenY: 0.37800002
   m_CameraDistance: 10
-  m_DeadZoneWidth: 0.358
-  m_DeadZoneHeight: 0.122
+  m_DeadZoneWidth: 0.248244
+  m_DeadZoneHeight: 0.1642265
   m_DeadZoneDepth: 1
   m_UnlimitedSoftZone: 0
-  m_SoftZoneWidth: 0.8
-  m_SoftZoneHeight: 0.8
-  m_BiasX: 0.076000236
-  m_BiasY: 0.3899995
+  m_SoftZoneWidth: 0.7651567
+  m_SoftZoneHeight: 1.171727
+  m_BiasX: 0.08112323
+  m_BiasY: 0.2461052
   m_GroupFramingMode: 2
   m_AdjustmentMode: 2
   m_GroupFramingSize: 0.8
@@ -33298,6 +33315,34 @@ CanvasRenderer:
   m_PrefabInternal: {fileID: 0}
   m_GameObject: {fileID: 1455786617}
   m_CullTransparentMesh: 0
+--- !u!1 &1472064909
+GameObject:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  serializedVersion: 6
+  m_Component:
+  - component: {fileID: 1472064910}
+  m_Layer: 0
+  m_Name: PlayerAvgPos
+  m_TagString: Untagged
+  m_Icon: {fileID: 0}
+  m_NavMeshLayer: 0
+  m_StaticEditorFlags: 0
+  m_IsActive: 1
+--- !u!4 &1472064910
+Transform:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInternal: {fileID: 0}
+  m_GameObject: {fileID: 1472064909}
+  m_LocalRotation: {x: -0, y: -0, z: -0, w: 1}
+  m_LocalPosition: {x: 0, y: 0, z: 0}
+  m_LocalScale: {x: 1, y: 1, z: 1}
+  m_Children: []
+  m_Father: {fileID: 814330893}
+  m_RootOrder: 3
+  m_LocalEulerAnglesHint: {x: 0, y: 0, z: 0}
 --- !u!1 &1481499263
 GameObject:
   m_ObjectHideFlags: 0
