+++ 3scape/Assets/Scripts/Camera/Score.cs
