+++ 3scape/Assets/Scripts/Camera/Timer.cs
